Add total recipe that using the tag

+++ src/main/java/com/cdcone/recipy/dtoAccess/TagDtoAdmin.java
@@ -9,5 +9,5 @@ public class TagDtoAdmin {
     private final int id;
     private final String name;
     private final Integer views;
-    private final Integer totalRecipe;
+    private final long totalRecipe;
 }
